Making new system for handling the alarm logic in the app
