+++ somnianalytics/somnianalytics/AlarmManager/AlarmService.swift
@@ -1,0 +1,14 @@
+//
+//  AlarmService.swift
+//  somnianalytics
+//
+//  Created by Hayden Barogh on 2/26/26.
+//
+
+import Foundation
+import AlarmKit
+
+@Observable
+final class AlarmService {
+    
+}
